공지등록 및 상세페이지조회 공지코드에 맞는 댓글 조회 완료 서식자료실 시작

+++ src/main/java/com/wisehr/wisehr/comment/entity/ComPosition.java
@@ -16,9 +16,9 @@ import lombok.*;
 public class ComPosition {
     @Id
     @Column(name = "pos_code", nullable = false)
-    private int posCode;
+    private Long posCode;
     @Column(name = "pos_name")
     private String posName;
     @Column(name = "pos_salary")
-    private int posSalary;
+    private Long posSalary;
 }
